move input resolver into transformer
