+++ src/core/jit/transform/input_resolver.rs
@@ -1,6 +1,6 @@
 use async_graphql_value::{ConstValue, Value};
 
-use super::{Arg, Field, OperationPlan, ResolveInputError, Variables};
+use super::super::{Arg, Field, OperationPlan, ResolveInputError, Variables};
 use crate::core::json::{JsonLikeOwned, JsonObjectLike};
 use crate::core::Type;
