+++ src/core/jit/transform/mod.rs
@@ -1,1 +1,5 @@
-pub mod skip;
+mod input_resolver;
+mod skip;
+
+pub use input_resolver::*;
+pub use skip::*;
